// apps/logistics/src/core/warehouse/inventory/application/dto/out/inventory-movement-response.dto.ts
export class InventoryMovementDetailResponseDto {
  id: number;
  productoNombre: string;
  cantidad: number;
  unidadMedida: string;
}

export class InventoryMovementResponseDto {
  id: number;
  tipoMovimiento: string;
  fechaMovimiento: Date;
  motivo: string;
  documentoReferencia: string;
  usuario: string;
  almacenOrigenNombre: string | null;
  almacenDestinoNombre: string | null;
  detalles: InventoryMovementDetailResponseDto[];
}